+++ changelog.md
@@ -3,17 +3,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## Early Dev [2018-10-07]
+
 ## Early Dev [2018-10-06]
+- HTTPS
+- Default ports switched.
+- Started Discord Webhook integration.
 - Suggestion Menu
 - Sorted by color properly now
 - Losing internet connection will display the proper message, and reconnect.
+- Pasting into the suggestion box wont paste the HTML
+- Database going down wont crash the server (probably)
 
-## Early Dev [2018-09-30]
+## Unreleased [2018-10-05]
 - Database on increasing numbers
 
-## Early Dev [2018-09-30]
+## Unreleased [2018-10-02]
 - Database Generation
 - Console Logs Look Nicer
 
-## Early Dev [2018-09-30]
+## Unreleased [2018-09-30]
 - Start Changelog